elevenlabs/requirements.txt
fastapi>=0.100.0,<0.115.0
httpx>=0.25.0,<0.28.0
uvicorn[standard]>=0.23.0,<0.30.0
python-dotenv>=1.0.0
boto3>=1.26.0